+++ cmd/state/exec3/state.go
@@ -21,8 +21,9 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/erigontech/erigon-lib/common/dbg"
 	"golang.org/x/sync/errgroup"
+
+	"github.com/erigontech/erigon-lib/common/dbg"
 
 	"github.com/erigontech/erigon-lib/log/v3"
