+++ txnprovider/txpool/pool_test.go
@@ -578,71 +578,6 @@ func TestTxnPoke(t *testing.T) {
 	}
 }
 
-func TestShanghaiIntrinsicGas(t *testing.T) {
-	cases := map[string]struct {
-		expected          uint64
-		dataLen           uint64
-		dataNonZeroLen    uint64
-		authorizationsLen uint64
-		creation          bool
-		isShanghai        bool
-	}{
-		"simple no data": {
-			expected:          21000,
-			dataLen:           0,
-			dataNonZeroLen:    0,
-			authorizationsLen: 0,
-			creation:          false,
-			isShanghai:        false,
-		},
-		"simple with data": {
-			expected:          21512,
-			dataLen:           32,
-			dataNonZeroLen:    32,
-			authorizationsLen: 0,
-			creation:          false,
-			isShanghai:        false,
-		},
-		"creation with data no shanghai": {
-			expected:          53512,
-			dataLen:           32,
-			dataNonZeroLen:    32,
-			authorizationsLen: 0,
-			creation:          true,
-			isShanghai:        false,
-		},
-		"creation with single word and shanghai": {
-			expected:          53514, // additional gas for single word
-			dataLen:           32,
-			dataNonZeroLen:    32,
-			authorizationsLen: 0,
-			creation:          true,
-			isShanghai:        true,
-		},
-		"creation between word 1 and 2 and shanghai": {
-			expected:          53532, // additional gas for going into 2nd word although not filling it
-			dataLen:           33,
-			dataNonZeroLen:    33,
-			authorizationsLen: 0,
-			creation:          true,
-			isShanghai:        true,
-		},
-	}
-
-	for name, c := range cases {
-		t.Run(name, func(t *testing.T) {
-			// Todo (@somnathb1) - Factor in EIP-7623
-			gas, _, reason := txpoolcfg.CalcIntrinsicGas(c.dataLen, c.dataNonZeroLen, c.authorizationsLen, nil, c.creation, true, true, c.isShanghai, false)
-			if reason != txpoolcfg.Success {
-				t.Errorf("expected success but got reason %v", reason)
-			}
-			if gas != c.expected {
-				t.Errorf("expected %v but got %v", c.expected, gas)
-			}
-		})
-	}
-}
-
 func TestShanghaiValidateTxn(t *testing.T) {
 	asrt := assert.New(t)
 	tests := map[string]struct {
